Breaking: Full support for websocket-based notifications and eol'ed http notifications
