+++ src/lib/db.js
@@ -17,8 +17,6 @@ const TABLE_NAMES = [
   'L_ACCOUNTS',
   'L_FULFILLMENTS',
   'L_ENTRIES',
-  'L_NOTIFICATIONS',
-  'L_SUBSCRIPTIONS',
   'L_TRANSFERS',
   'L_LU_REJECTION_REASON',
   'L_LU_TRANSFER_STATUS'
