+++ src/services/app.js
@@ -5,5 +5,5 @@ module.exports = new App({
   log: require('./log'),
   config: require('./config'),
   timerWorker: require('./timerWorker'),
-  notificationWorker: require('./notificationWorker')
+  notificationBroadcaster: require('./notificationBroadcaster')
 })
