+++ src/services/notificationBroadcaster.js
@@ -1,0 +1,6 @@
+'use strict'
+
+const NotificationBroadcaster = require('../lib/notificationBroadcasterWebsocket')
+const log = require('./log')
+
+module.exports = new NotificationBroadcaster(log('notificationBroadcaster'))
