--- src/services/notificationWorker.js
@@ -1,8 +1,0 @@
-'use strict'
-
-const NotificationWorker = require('../lib/notificationWorker')
-const uri = require('./uriManager')
-const log = require('./log')
-const config = require('./config')
-
-module.exports = new NotificationWorker(uri, log('notificationWorker'), config)
