+++ src/services/transferExpiryMonitor.js
@@ -3,6 +3,6 @@
 const TransferExpiryMonitor =
   require('../lib/transferExpiryMonitor').TransferExpiryMonitor
 const timeQueue = require('./timeQueue')
-const notificationWorker = require('./notificationWorker')
+const notificationBroadcaster = require('./notificationBroadcaster')
 
-module.exports = new TransferExpiryMonitor(timeQueue, notificationWorker)
+module.exports = new TransferExpiryMonitor(timeQueue, notificationBroadcaster)
