+++ src/sql/pg/create.sql
@@ -87,7 +87,6 @@ CREATE INDEX "L_XIF_TRANSFERS_STATE" ON "L_TRANSFERS"
 CREATE INDEX "L_XIF_TRANSFERS_REASON" ON "L_TRANSFERS"
   ("REJECTION_REASON_ID" ASC);
 
-
 CREATE TABLE "L_TRANSFER_ADJUSTMENTS"
 (
   "TRANSFER_ADJUSTMENT_ID" SERIAL NOT NULL,
@@ -111,58 +110,6 @@ CREATE INDEX "L_XIF_TRANSFER_ADJUSTMENTS_ACC" ON "L_TRANSFER_ADJUSTMENTS"
   ("ACCOUNT_ID" ASC);
 CREATE INDEX "L_XIE_TRANSFER_ADJUSTMENTS" ON "L_TRANSFER_ADJUSTMENTS"
   ("IS_AUTHORIZED" ASC);
-
-
-CREATE TABLE IF NOT EXISTS "L_SUBSCRIPTIONS" (
-  "SUBSCRIPTION_ID" CHARACTER VARYING(36) NOT NULL,
-  "OWNER" CHARACTER VARYING(255) NOT NULL,
-  "EVENT" CHARACTER VARYING(255) NOT NULL,
-  "SUBJECT" CHARACTER VARYING(1024) NOT NULL,
-  "TARGET" CHARACTER VARYING(1024) NOT NULL,
-  "IS_DELETED" BOOLEAN NOT NULL DEFAULT FALSE
-);
-
-CREATE INDEX "L_XPK_SUBSCRIPTIONS" ON "L_SUBSCRIPTIONS"
-  ("SUBSCRIPTION_ID" ASC);
-ALTER TABLE "L_SUBSCRIPTIONS" ADD CONSTRAINT "L_PK_SUBSCRIPTIONS" PRIMARY KEY
-  ("SUBSCRIPTION_ID");
--- CREATE UNIQUE INDEX "L_XAK_SUBSCRIPTIONS" ON "L_SUBSCRIPTIONS"
---   ("SUBSCRIPTION_UUID" ASC);
--- ALTER TABLE "L_SUBSCRIPTIONS" ADD CONSTRAINT "L_XAK_SUBSCRIPTIONS" UNIQUE
---   ("SUBSCRIPTION_UUID");
-CREATE INDEX "L_XIF_SUBSCRIPTIONS_OWNER" ON "L_SUBSCRIPTIONS"
-  ("OWNER" ASC);
-CREATE INDEX "L_XIF_SUBSCRIPTIONS_SUBJECT" ON "L_SUBSCRIPTIONS"
-  ("SUBJECT" ASC);
-CREATE INDEX "L_XIE_SUBSCRIPTIONS_DELETED" ON "L_SUBSCRIPTIONS"
-  ("IS_DELETED" ASC);
-
-
-CREATE TABLE IF NOT EXISTS "L_NOTIFICATIONS" (
-  "NOTIFICATION_ID" CHARACTER(36) NOT NULL,
-  "SUBSCRIPTION_ID" CHARACTER(36) NOT NULL,
-  "TRANSFER_ID" CHARACTER(36) NOT NULL,
-  "RETRY_COUNT" INTEGER DEFAULT 0 NOT NULL,
-  "RETRY_AT" TIMESTAMP WITH TIME ZONE NULL
-);
-
-CREATE INDEX "L_XPK_NOTIFICATIONS" ON "L_NOTIFICATIONS"
-  ("NOTIFICATION_ID" ASC);
-ALTER TABLE "L_NOTIFICATIONS" ADD CONSTRAINT "L_PK_NOTIFICATIONS" PRIMARY KEY
-  ("NOTIFICATION_ID");
--- CREATE INDEX "L_XAK_NOTIFICATIONS" ON "L_NOTIFICATIONS"
---   ("NOTIFICATION_UUID" ASC);
--- ALTER TABLE "L_NOTIFICATIONS" ADD CONSTRAINT "L_AK_NOTIFICATIONS" UNIQUE
---   ("NOTIFICATION_UUID");
-CREATE INDEX "L_XIE_NOTIFICATIONS_RETRY_AT" ON "L_NOTIFICATIONS"
-  ("RETRY_AT" ASC);
-CREATE INDEX "L_XIF_NOTIFICATIONS_SUB" ON "L_NOTIFICATIONS"
-  ("SUBSCRIPTION_ID" ASC);
-CREATE INDEX "L_XIF_NOTIFICATIONS_TRANSFER" ON "L_NOTIFICATIONS"
-  ("TRANSFER_ID" ASC);
--- CREATE INDEX "L_XIE_NOTIFICATIONS_DELETED" ON "L_NOTIFICATIONS"
---   ("IS_DELETED" ASC);
-
 
 CREATE TABLE IF NOT EXISTS "L_ENTRIES" (
   "ENTRY_ID" SERIAL NOT NULL,
