+++ src/sql/pg/drop.sql
@@ -2,8 +2,6 @@ DROP TABLE IF EXISTS "L_TRANSFER_ADJUSTMENTS" CASCADE;
 DROP TABLE IF EXISTS "L_ACCOUNTS" CASCADE;
 DROP TABLE IF EXISTS "L_LU_REJECTION_REASON" CASCADE;
 DROP TABLE IF EXISTS "L_LU_TRANSFER_STATUS" CASCADE;
-DROP TABLE IF EXISTS "L_SUBSCRIPTIONS" CASCADE;
 DROP TABLE IF EXISTS "L_ENTRIES" CASCADE;
 DROP TABLE IF EXISTS "L_FULFILLMENTS" CASCADE;
-DROP TABLE IF EXISTS "L_NOTIFICATIONS" CASCADE;
 DROP TABLE IF EXISTS "L_TRANSFERS" CASCADE;
