+++ src/sql/sqlite3/drop.sql
@@ -2,8 +2,6 @@ DROP TABLE IF EXISTS L_TRANSFER_ADJUSTMENTS;
 DROP TABLE IF EXISTS L_ACCOUNTS;
 DROP TABLE IF EXISTS L_LU_REJECTION_REASON;
 DROP TABLE IF EXISTS L_LU_TRANSFER_STATUS;
-DROP TABLE IF EXISTS L_SUBSCRIPTIONS;
 DROP TABLE IF EXISTS L_ENTRIES;
 DROP TABLE IF EXISTS L_FULFILLMENTS;
-DROP TABLE IF EXISTS L_NOTIFICATIONS;
 DROP TABLE IF EXISTS L_TRANSFERS;
