+++ src/sql/strong-oracle/create.sql
@@ -66,30 +66,6 @@ CREATE INDEX L_XIE_FINGERPRINTS ON "L_ACCOUNTS"
   ("FINGERPRINT" ASC)
 /
 
-
-CREATE TABLE "L_SUBSCRIPTIONS"
-(
-  "SUBSCRIPTION_ID"      VARCHAR2(64) NOT NULL ,
-  "OWNER"                VARCHAR2(1024) NULL ,
-  "EVENT"                VARCHAR2(255) DEFAULT NULL NULL ,
-  "SUBJECT"              VARCHAR2(1024) NULL ,
-  "TARGET"               VARCHAR2(1024) DEFAULT NULL NULL ,
-  "IS_DELETED"           SMALLINT DEFAULT 0 NOT NULL
-)
-/
-
-CREATE INDEX L_XPK_SUBSCRIPTIONS ON "L_SUBSCRIPTIONS"
-  ("SUBSCRIPTION_ID" ASC)
-/
-
-ALTER TABLE "L_SUBSCRIPTIONS" ADD CONSTRAINT L_PK_SUBSCRIPTIONS PRIMARY KEY
-  ("SUBSCRIPTION_ID")
-/
-
-CREATE INDEX L_XIE_SUBSCRIPTIONS ON "L_SUBSCRIPTIONS"
-  ("IS_DELETED" ASC)
-/
-
 CREATE TABLE "L_LU_REJECTION_REASON" (
   "REJECTION_REASON_ID"   INTEGER NOT NULL,
   "NAME"                  CHARACTER VARYING(10) NOT NULL,
@@ -257,36 +233,6 @@ ALTER TABLE "L_FULFILLMENTS" ADD CONSTRAINT L_PK_FULFILLMENTS PRIMARY KEY
 /
 
 CREATE INDEX L_XIF_FULFILLMENTS ON "L_FULFILLMENTS"
-  ("TRANSFER_ID" ASC)
-/
-
-CREATE TABLE "L_NOTIFICATIONS"
-(
-  "NOTIFICATION_ID"      VARCHAR2(36) NOT NULL ,
-  "SUBSCRIPTION_ID"      VARCHAR2(36) NULL ,
-  "TRANSFER_ID"          VARCHAR2(36) NULL ,
-  "RETRY_COUNT"          INTEGER DEFAULT 0 NULL ,
-  "RETRY_AT"             TIMESTAMP WITH TIME ZONE
-)
-/
-
-CREATE INDEX XPKL_NOTIFICATIONS ON "L_NOTIFICATIONS"
-  ("NOTIFICATION_ID" ASC)
-/
-
-ALTER TABLE "L_NOTIFICATIONS" ADD CONSTRAINT L_PK_NOTIFICATIONS PRIMARY KEY
-  ("NOTIFICATION_ID")
-/
-
-CREATE INDEX L_XIE_NOTIFICATIONS_RETRY_AT ON "L_NOTIFICATIONS"
-  ("RETRY_AT" ASC)
-/
-
-CREATE INDEX L_XIF_NOTIFICATIONS_SUB ON "L_NOTIFICATIONS"
-  ("SUBSCRIPTION_ID" ASC)
-/
-
-CREATE INDEX L_XIF_NOTIFICATIONS_TRANSFER ON "L_NOTIFICATIONS"
   ("TRANSFER_ID" ASC)
 /
