+++ src/sql/strong-oracle/drop.sql
@@ -39,16 +39,6 @@ END;
 /
 
 BEGIN
-   EXECUTE IMMEDIATE 'DROP TABLE "L_SUBSCRIPTIONS"';
-EXCEPTION
-   WHEN OTHERS THEN
-      IF SQLCODE != -942 THEN
-         RAISE;
-      END IF;
-END;
-/
-
-BEGIN
    EXECUTE IMMEDIATE 'DROP TABLE "L_ENTRIES"';
 EXCEPTION
    WHEN OTHERS THEN
@@ -70,16 +60,6 @@ END;
 
 BEGIN
    EXECUTE IMMEDIATE 'DROP TABLE "L_FULFILLMENTS"';
-EXCEPTION
-   WHEN OTHERS THEN
-      IF SQLCODE != -942 THEN
-         RAISE;
-      END IF;
-END;
-/
-
-BEGIN
-   EXECUTE IMMEDIATE 'DROP TABLE "L_NOTIFICATIONS"';
 EXCEPTION
    WHEN OTHERS THEN
       IF SQLCODE != -942 THEN
